When the ComponentUI for a component is not found, it can come from the well-known "ClassLoader" problem where UIManager is not able to find the ComponentUI class because it uses different classloader OR it can be that the ComponentUI is simply not registered and UIManager.get("xxxxUI") returns null.
In the case of Applet, the latter applies. Whenever you refresh an applet, it
seems the internal Swing "appContext" (in SwingUtilities) is recreated. It is
equivalent to reinitializing UIManager - all previous customizations are lost.
This should not be a problem for us except when you refresh the applet, it seems
classes do not get reloaded thus the "static" code we use in say JXTaskPane to
register UIManager settings does not get executed. Later a call to
UIManager.get("TaskPaneUI") will return null.

There is already some workaround in LookAndFeelAddons for environment such as
IDEs in LookAndFeelAddons#getUI(). This is the place where we can also add a
check. This check should make sure the current addon has registered its defaults.

Issue 142