started fix on #459-swingx: table.setLocale doesn't update locale-dependent column action state
- LookAndFeelAddon: added "old" handling of swingx resource bundle as comment. We need to explore to more depth how to allow access to
resource values in different locales. For now (and as the app resourceBundle isn't working reliable anyway in jdk5) interested parties
need to out-comment.

- JXTable: updates the action's locale dependent state by accessing the
UIManager's default by the method taking the current locale. This update
is called in setLocale after super. Note, that the fix does work only if
a) the resourceBundle is added to the uiManager (changed LookAndFeelAddon above)
AND
b) jdk6 is used.

- updated tests